Made the normalizing flows code more general to avoid repetition when used for background.
